pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssafer"
version = "0.1.0"
description = "Windows-first security configuration artifact collector for SSAfer."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.0",
  "rich>=13.7.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
ssafer = "ssafer.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
